Added a bunch of functions: Calculate average hamming distance in train matches; Calculate hamming distance of matches with video descriptors; Filter out matches that have distance large than hamming distance.

+++ featureMatching/functions.cpp
@@ -128,7 +128,7 @@ void writeDescriptorsToFile(vector<ImageFrame>* images){
     
     /* Print number of images/descriptors in the file. */
     
-    int numDescriptors = images->size();
+    long numDescriptors = images->size();
     
     outputFile << numDescriptors << endl << endl;
